==> notebooks/01. extraccion_gmaps.ipynb <==
# Importar librerias

import pandas as pd
import json
import os

# Extracción

Se carga y se cambia de formato el dataset crudo para que sea mas sencillo cargar y limpiar mas adelante.

## Metadata de hoteles

# Ruta de la carpeta de datos de Google Maps
google_maps_folder = '../raw_data/Google Maps/'

# Ruta de la subcarpeta "metadata-sitios"
google_maps_subfolder = os.path.join(google_maps_folder, 'metadata-sitios')

# Lista para almacenar los resultados filtrados
hotels = []

# Recorrer todos los archivos JSON dentro de "metadata-sitios"
for archivo in os.listdir(google_maps_subfolder):
    if archivo.endswith(".json"):
        # Ruta del archivo JSON actual
        ruta_archivo = os.path.join(google_maps_subfolder, archivo)

        # Leer el contenido del archivo JSON
        with open(ruta_archivo) as f:
            contenido_json = f.readlines()

        # Cargar cada objeto JSON por separado
        for linea in contenido_json:
            try:
                data = json.loads(linea)
            except json.JSONDecodeError as e:
                print(f"Error al cargar el archivo JSON {ruta_archivo}: {e}")
                continue

            # Verificar si la categoría "hotel" está presente en el archivo JSON
            categorias = data.get("category", [])
            if isinstance(categorias, list) and any("hotel" in categoria.lower() for categoria in categorias):
                hotels.append(data)

# Convertir los resultados a un pandas DataFrame
dfHotels = pd.DataFrame(hotels)

dfHotels.to_csv('../processed_data/hoteles_crudo.csv', index=False, sep='|')

dfHotels.info()

dfHotels.head()

print('Duplicados "hotels":', dfHotels['gmap_id'].duplicated().sum())

dfHotels = dfHotels.drop_duplicates(subset='gmap_id') #63 observaciones

dfHotels.shape

# Reviews por estado

## Funcion para listar subcarpetas

def get_subfolders(folder_path):
    subfolders = [f.path for f in os.scandir(folder_path) if f.is_dir()]
    return subfolders

folder_path_reviews_estados = "../raw_data/Google Maps/reviews-estados"
subfolders = get_subfolders(folder_path_reviews_estados)

subfolders

## Funcion para unir archivos json

def read_json_files(folder_path):
    data = []
    
    for file_name in os.listdir(folder_path):
        if file_name.endswith(".json"):
            file_path = os.path.join(folder_path, file_name)
            with open(file_path, 'r') as file:
                for line in file:
                    line = line.strip()
                    if line:
                        try:
                            json_obj = json.loads(line)
                            data.append(json_obj)
                        except json.JSONDecodeError:
                            continue

    df = pd.DataFrame(data)
    return df

## Funcion para unir los json en cada subcarpeta

Dependiendo de la capacidad de computo, es posible que se necesite realizar este paso con pocos archivos a la vez.

def read_json_files_in_subfolders(subfolders):
    df_list = []
    
    for subfolder in subfolders:
        df = read_json_files(subfolder)
        df_list.append(df)
    
    return df_list

dataframes_list = read_json_files_in_subfolders(subfolders)

Concatenar la lista en un df y guardar.

dfreviews = pd.concat(dataframes_list, ignore_index=True)

dfreviews.to_parquet('../raw_data/Google Maps/reviews-estados/reviews_google_crudo.parquet', index=False, compression="gzip")

En caso tal de haber dividido los archivos en varios parquet, se puede usar este codigo para concatenarlos y luego hacer merge con el dataframe de hoteles para reducir su tamaño.

def load_and_concat_parquet_files(folder_path):
    # Obtener una lista de todos los archivos Parquet en la carpeta
    parquet_files = [file for file in os.listdir(folder_path) if file.endswith('.parquet')]

    # Inicializar una lista para almacenar los DataFrames
    dfs = []

    # Cargar cada archivo Parquet en un DataFrame y agregarlo a la lista
    for file in parquet_files:
        file_path = os.path.join(folder_path, file)
        df = pd.read_parquet(file_path)
        dfs.append(df)

    # Concatenar los DataFrames en uno solo
    concatenated_df = pd.concat(dfs, ignore_index=True)

    return concatenated_df

folder_path = '../raw_data/Google Maps/reviews-estados/'
reviews_df = load_and_concat_parquet_files(folder_path)


## Uniendo hoteles con reviews

inner_merged = pd.merge(reviews_df, dfHotels, on=["gmap_id"])

inner_merged.to_csv("../raw_data/Google Maps/reviews-estados/reviews_hoteles_crudo.csv", index=False,  sep='|')
