==> notebooks/02. transformacion_gmaps.ipynb <==
# Importar librerias y dataset

import pandas as pd
import re
import warnings
warnings.filterwarnings("ignore")

hoteles = pd.read_csv("../raw_data/Google Maps/reviews-estados/hoteles_crudo.csv", sep="|")

reviews = pd.read_csv("../raw_data/Google Maps/reviews-estados/reviews_hoteles_crudo.csv", sep="|")

# Nulos y duplicados

## Hoteles

hoteles.isna().mean()

borrar_columnas = ["description", "price", "hours", "MISC", "relative_results", "url"]
hoteles.drop(columns = borrar_columnas, inplace = True)

hoteles.dropna(subset = ["name", "address"], inplace=True)

hoteles.dtypes

hoteles.shape

hoteles.duplicated(subset='name').sum()

hoteles.drop_duplicates(subset='name', inplace=True) 

## Reviews

reviews.isna().mean()

borrar_columnas = ["pics", "resp"]
reviews.drop(columns=borrar_columnas, inplace=True)

reviews.dropna(subset = ["text"], inplace=True)

reviews.duplicated().sum()

reviews.drop_duplicates(inplace=True) 

reviews.dtypes

reviews.shape

# Normalización

## Tablas dimensionales

### Business, category y state

hoteles.head()

business = hoteles.reset_index(drop=True)

business = business.reset_index()
business.rename(columns={'index': 'business_id', 'name': 'business_name'}, inplace=True)
business.head()

valid_state_codes = ["AL", "AK", "AZ", "AR", "CA", "CO", "CT", "DC", "DE", "FL", "GA", "HI", "ID", "IL", "IN", "IA", "KS", 
                    "KY", "LA", "ME", "MD", "MA", "MI", "MN", "MS", "MO", "MT", "NE", "NV", "NH", "NJ", "NM", "NY", 
                    "NC", "ND", "OH", "OK", "OR", "PA", "RI", "SC", "SD", "TN", "TX", "UT", "VT", "VA", "WA", "WV", "WI", "WY"]

def extract_state_code(address):
    # Expresión regular para buscar códigos de estado válidos
    regex_pattern = r"\b(" + "|".join(valid_state_codes) + r")\b"

    # Busca el código de estado en la dirección usando regex
    match = re.search(regex_pattern, address)

    # Devuelve el código de estado si se encuentra, o None si no se encuentra
    return match.group(0) if match else None

# Crea una nueva columna "state" aplicando la función a la columna "address"

business['state'] = business['address'].apply(extract_state_code)

business.isna().sum()

business.dropna(subset="state", inplace=True)

business.head()

Usamos explode() para expandir las listas de la columna category, los nuevos registros de categoria conservan su respectivo index de la anterior tabla, pero al haber creado la columna business_id con valores iguales al index de dicha tabla, los valores de categoria ahora coinciden correctamente con el id de los hoteles.

import ast

business['category'] = business['category'].apply(ast.literal_eval)

business_category = business.explode('category')["category"].reset_index()
business_category.rename(columns={'index': 'business_id'}, inplace=True)
business_category.head()

Ahora con los valores unicos se puede crear la tabla de categoria, hacer un join posteriormente y borrar la columna category de business.

category = business_category["category"].drop_duplicates().reset_index(drop=True) 
category = category.reset_index()
category.rename(columns={'index': 'category_id'}, inplace=True)
category

business_category = business_category.merge(category, on="category").sort_values(by="business_id").drop(columns="category").reset_index(drop=True)
business_category.head()

state_names = ["Alabama", "Alaska", "Arizona", "Arkansas", "California", "Colorado", "Connecticut", "Washington DC", "Delaware", "Florida", "Georgia",
               "Hawaii", "Idaho", "Illinois", "Indiana", "Iowa", "Kansas", "Kentucky", "Louisiana", "Maine", "Maryland", "Massachusetts",
               "Michigan", "Minnesota", "Mississippi", "Missouri", "Montana", "Nebraska", "Nevada", "New Hampshire", "New Jersey", "New Mexico",
               "New York", "North Carolina", "North Dakota", "Ohio", "Oklahoma", "Oregon", "Pennsylvania", "Rhode Island", "South Carolina",
               "South Dakota", "Tennessee", "Texas", "Utah", "Vermont", "Virginia", "Washington", "West Virginia", "Wisconsin", "Wyoming"]

states = pd.DataFrame()
states["state"] = valid_state_codes
states["state_name"] = state_names

### Users

reviews.head()

users = reviews[["user_id", "name_x"]].drop_duplicates(subset=['user_id', 'name_x']).reset_index(drop=True)

users.duplicated(subset="user_id").sum()

users.drop_duplicates(subset="user_id", inplace=True)

users.rename(columns={'user_id': 'google_id'}, inplace=True)

users = users.reset_index(drop=True)

users = users.reset_index()
users.rename(columns={'index': 'user_id', 'name_x': 'user_name'}, inplace=True)
users

## Tabla de hechos

reviews['time'] = pd.to_datetime(reviews['time'], unit='ms')
reviews['time'] = reviews['time'].dt.strftime('%Y-%m-%d')

reviews.drop(columns="name_x", inplace=True)

reviews.rename(columns={'time': 'date', 'user_id': 'google_id',
                        'text': 'review'}, inplace=True)

reviews = reviews.sort_values(by="date").reset_index(drop=True)

reviews = reviews.reset_index() 
reviews.rename(columns={'index': 'review_id'}, inplace=True)
reviews

## Guardando dataset

# Exportamos
reviews.to_csv('../processed_data/reviews_google_normalizado.csv', index=False, sep='|')

users.to_csv('../processed_data/users_google.csv', index=False, sep='|')

business.to_csv('../processed_data/business_google.csv', index=False, sep='|')

category.to_csv('../processed_data/category_google.csv', index=False, sep='|')

business_category.to_csv('../processed_data/business_category_google.csv', index=False, sep='|')

states.to_csv('../processed_data/states.csv', index=False, sep='|')

Para efectos practicos de exploracion de datos y modelado, se guarda tambien una tabla no normalizada.

reviews_merge = reviews.merge(business, on="gmap_id")

reviews_merge.to_csv('../processed_data/reviews_google_sin_normalizar.csv', index=False, sep='|')
