==> notebooks/04. etl_yelp.ipynb <==
# Yelp dataset cleaning

#Cargamos las librerías necesarias para hacer la limpieza
import pandas as pd
import os

# Asignamos el path relativo a la variable dir para utilizarla a la hora de consumir los datasets
# Obtener el path actual
dir = os.getcwd()

# Cargamos el dataset 'business.parquet' que contiene la lista de negocios
df_yelp_business = pd.read_parquet(dir+'/parquets/business.parquet')


df_yelp_business.head()

df_yelp_business.info()

### Extraemos las categorías del dataset

df = pd.DataFrame()
# Convertir los strings separados por comas en una lista y obtener los valores únicos
df['categories_list'] = df_yelp_business['categories'].str.split(', ')
unique_categories = set(df['categories_list'].explode())
# Convertir el conjunto a una lista
lista_unique_categories = list(unique_categories)

## Creamos una tabla con filtro en la columna 'categories'

df_hotels_yelp = df_yelp_business
# Primero, dividimos los valores de 'categories' en listas de strings
df_hotels_yelp['categories'] = df_hotels_yelp['categories'].str.split(',')

df_hotels_yelp = df_hotels_yelp.dropna(subset=['categories'])

# A continuación, definimos una función que verifique si 'hotels' aparece en la lista,
# y si 'hotels & travel' o 'hotel bar' están ausentes
def filter_categories(categories):
    return 'Hotels' in categories and 'Hotels & Travel' not in categories and 'Hotel Bar' not in categories

# Usamos la función para filtrar las filas y nos quedamos solo con las que cumplen el criterio
df_hotels_yelp = df_hotels_yelp[df_hotels_yelp['categories'].apply(filter_categories)]

# Ahora, si lo deseas, puedes volver a unir los valores de la lista en un solo string separado por comas
df_hotels_yelp['categories'] = df_hotels_yelp['categories'].apply(lambda x: ','.join(x))

# Restablecemos el índice y eliminar la columna "index"
df_hotels_yelp.reset_index(inplace=True, drop=True)

df_hotels_yelp.head()

df_hotels_yelp.shape

df_hotels_yelp.info()

### Filtro el df de hoteles solo en los estados 'CA', 'FL', 'NV', 'NY' y 'TX'

# Filtrar filas donde los valores de la columna 'state' los contanga la siguiente lista: ['CA', 'FL', 'NV', 'NY', 'TX'] 
df_hotels_yelp = df_hotels_yelp[df_hotels_yelp['state'].isin(['CA', 'FL', 'NV', 'NY', 'TX'])]

# Resetear indices
df_hotels_yelp.reset_index(inplace=True,drop=True)


df_hotels_yelp.shape

df_hotels_yelp.to_csv('hotels_yelp.csv')

### Creo la tabla yelp_hotel_categories que contiene las categorias y sus id

df = pd.DataFrame()
# Convertir los strings separados por comas en una lista y obtener los valores únicos
df['categories_list'] = df_hotels_yelp['categories'].str.split(', ')
unique_hotels_categories = set(df['categories_list'].explode())
# Convertir el conjunto a una lista
lista_hotels_categories = list(unique_hotels_categories)
# Create a DataFrame with columns 'category_id' and 'category'
lista_hotels_categories = pd.DataFrame({'category_id': range(len(lista_hotels_categories)), 'category': lista_hotels_categories})

lista_hotels_categories.reset_index(inplace=True,drop=True)

lista_hotels_categories.to_csv('hotel_categories_yelp.csv', index=False)

# Transformo el dataframe en un diccionario para luego mapear el dataframe 'df_hotels_desanidado'
category_dict = lista_hotels_categories.set_index('category')['category_id'].to_dict()

category_dict

### Creo una tabla que contenga los business_id y los categories_id

#Primero desanido la tabla df_hotels_yelp
df_anidado = df_hotels_yelp.loc[:,['business_id','categories']]

# Dividir los valores separados por comas en filas separadas
df_anidado['categories'] = df_anidado['categories'].str.split(',')

# Usar la función explode para obtener un DataFrame desanidado
df_hotels_desanidado = df_anidado.explode('categories')

#Elimino los espacios en los valores de la columna 'categories'
df_hotels_desanidado['categories'] = df_hotels_desanidado['categories'].str.strip()

# Renombrar la columna 'categories'
df_hotels_desanidado.rename(columns={'categories': 'category'}, inplace=True)

# Elimino los index
df_hotels_desanidado.reset_index(inplace=True, drop=True)

# Mostrar el nuevo DataFrame desanidado
df_hotels_desanidado.head()



df_hotels_desanidado.shape

### Creo una tabla que contenga 'business_id', 'category' y 'category_id'

business_category_yelp = df_hotels_desanidado

# Para agregar la columna 'category_id' al DataFrame, puedes usar el método map:
business_category_yelp['category_id'] = business_category_yelp['category'].map(category_dict)

# Borro la columna 'category'
del business_category_yelp["category"]

business_category_yelp.head()


business_category_yelp.to_csv('business_category_yelp.csv', index=False)

### Creo una tabla con las columnas 'business_id' del dataframe 'df_hotels_desanidado' y la columna 'category_id' del dataframe 'lista_hotels_categories'


business_category = lista_hotels_categories.merge(df_hotels_desanidado, on="category").sort_values(by="business_id").drop(columns="category").reset_index(drop=True)
business_category.head()

### Creo una tabla que tenga los nombres de los hoteles

df_hotels_name = df_hotels_yelp['name']
df_hotels_name.head()


df_hotels_name.to_csv('hotels_name_yelp.csv',index=False)

### Creo tabla de palabras frecuentes en los nombres

# Obtener una Serie con todas las palabras de la columna 'nombre_columna'
palabras = df_hotels_yelp['name'].str.split(expand=True).stack()

# Contar la frecuencia de cada palabra y ordenarlas de mayor a menor frecuencia
frecuencia_palabras = palabras.value_counts().reset_index()
frecuencia_palabras.columns = ['Palabra', 'Frecuencia']
frecuencia_palabras = frecuencia_palabras.sort_values(by='Frecuencia', ascending=False)

print(frecuencia_palabras)

frecuencia_palabras.to_csv('palabras_frecuencia_name_yelp.csv',index=False)

## Trabajamos sobre el dataset checkin

df_yelp_checkin = pd.read_parquet(dir+'/parquets/checkin.parquet')


df_yelp_checkin.head()

# Filtrar filas de df_yelp_checkin donde 'business_id' coincida con los valores de df_hotels_name
yelp_checkin = df_yelp_checkin[df_yelp_checkin['business_id'].isin(df_hotels_yelp['business_id'])]

yelp_checkin.reset_index(inplace=True,drop=True)


yelp_checkin.info()

yelp_checkin.to_csv('checkin_yelp.csv',index=False)

# Trabajamos sobre el dataset review

df_yelp_review = pd.read_parquet(dir+'/parquets/review.parquet')


df_yelp_review.head()

df_yelp_review.info()

# Creamos una tabla 'review_merge' con las filas filtradas según los valores de la columna 'business_id' de df_final
# Obtener los valores únicos de la columna "business_id" del DataFrame df_final
valores_filtro = df_hotels_yelp['business_id'].unique()

# Filtrar las filas de review_merge basado en los valores de "business_id" que coincidan con "business_id" en df_final
yelp_review = df_yelp_review[df_yelp_review['business_id'].isin(valores_filtro)]

# Restablecemos el índice y eliminar la columna "index"
yelp_review.reset_index(inplace=True, drop=True)

yelp_review.head()

yelp_review.info()

yelp_review.to_csv('review_yelp.csv',index=False)

## Trabajamos con el dataset tip

df_yelp_tip = pd.read_parquet(dir+'/parquets/tip.parquet')


df_yelp_tip.head()

df_yelp_tip.shape

# Creamos una tabla 'tip_merge' con las filas filtradas según los valores de la columna 'business_id' de df_final
# Obtener los valores únicos de la columna "business_id" del DataFrame df_final
valores_filtro = df_hotels_yelp['business_id'].unique()

# Filtrar las filas de review_merge basado en los valores de "business_id" que coincidan con "business_id" en df_final
yelp_tip = df_yelp_tip[df_yelp_tip['business_id'].isin(valores_filtro)]

# Restablecemos el índice y eliminar la columna "index"
yelp_tip.reset_index(inplace=True, drop=True)

yelp_tip.head()

yelp_tip.shape

yelp_tip.to_csv('tip_yelp.csv',index=False)

## Trabajamos sobre la tabla user

df_yelp_user = pd.read_parquet(dir+'/user.parquet')

df_yelp_user.head()

df_yelp_user.shape

### Creo una tabla de usuarios que coinciden con las tablas review y otra con tip para futuro posible uso

### Tabla que contiene los usuarios que dejaron reviews a los hoteles del dataset

# Creo tabla 'user_review' que es la tabla 'user' filtrada con los 'user_id' de la tabla 'review_merge'
# Obtener los valores únicos de la columna "business_id" del DataFrame df_final
valores_filtro = yelp_review['user_id'].unique()

# Filtrar las filas de review_merge basado en los valores de "business_id" que coincidan con "business_id" en df_final
yelp_user_review = df_yelp_user[df_yelp_user['user_id'].isin(valores_filtro)]

# Restablecemos el índice y eliminar la columna "index"
yelp_user_review.reset_index(inplace=True, drop=True)

yelp_user_review.head()

yelp_user_review.shape

yelp_user_review.info()

yelp_user_review.to_csv('user_review_yelp.csv',index=False)

### Tabla que contiene los usuarios que dejaron tips a los hoteles del dataset

# Creo tabla 'user_tip' que es la tabla 'user' filtrada con los 'user_id' de la tabla 'tip_merge'
# Obtener los valores únicos de la columna "business_id" del DataFrame df_final
valores_filtro = yelp_tip['user_id'].unique()

# Filtrar las filas de review_merge basado en los valores de "business_id" que coincidan con "business_id" en df_final
yelp_user_tip = df_yelp_user[df_yelp_user['user_id'].isin(valores_filtro)]

# Restablecemos el índice y eliminar la columna "index"
yelp_user_tip.reset_index(inplace=True, drop=True)

yelp_user_tip.head()

yelp_user_tip.shape

yelp_user_tip.to_csv('user_tip_yelp.csv',index=False)
