==> notebooks/10. etl_ramada.ipynb <==
# Importar librerias y data

import pandas as pd

ramada_houston = pd.read_excel("/content/drive/MyDrive/Colab Notebooks/hoteles-ramada/ramada-houston-texas.xlsx", parse_dates=["review_datetime_utc"])
ramada_losangeles = pd.read_excel("/content/drive/MyDrive/Colab Notebooks/hoteles-ramada/ramada-los-angeles-california.xlsx", parse_dates=["review_datetime_utc"])
ramada_orlando = pd.read_excel("/content/drive/MyDrive/Colab Notebooks/hoteles-ramada/ramada-orlando-florida.xlsx", parse_dates=["review_datetime_utc"])
ramada_newyork = pd.read_excel("/content/drive/MyDrive/Colab Notebooks/hoteles-ramada/ramada-newyork-newyork.xlsx", parse_dates=["review_datetime_utc"])
ramada_reno = pd.read_excel("/content/drive/MyDrive/Colab Notebooks/hoteles-ramada/ramada-reno-nevada.xlsx", parse_dates=["review_datetime_utc"])

# Houston

ramada_houston.shape

ramada_houston.isna().mean()

drop_columns = ["query", "location_link", "google_id", "reviews_link", "reviews", "rating", "reviews_per_score", "author_link",
                "author_image", "review_img_url", "review_img_urls", "owner_answer", "owner_answer_timestamp",
                "owner_answer_timestamp_datetime_utc", "review_link", "review_timestamp", "review_likes", "reviews_id"]

ramada_houston.drop(columns=drop_columns, inplace=True)

ramada_houston['review_datetime_utc'] = ramada_houston['review_datetime_utc'].dt.date

ramada_houston["state"] = "TX"

ramada_houston.rename(columns={"name":"hotel_name",
                   "author_title": "user_name", "reviews": "num_of_reviews",
                   "author_id": "user_id", "review_text": "review", "review_rating": "rating",
                   "review_datetime_utc": "date"}, inplace=True)

ramada_houston.isna().mean()

# Resto de estados

def etl(df):
  drop_columns = ["query", "location_link",  "google_id", "reviews_link", "reviews", "rating", "reviews_per_score", "author_link",
                "author_image", "review_img_url", "review_img_urls", "owner_answer", "owner_answer_timestamp",
                "owner_answer_timestamp_datetime_utc", "review_link", "review_timestamp", "review_likes", "reviews_id"]

  df.drop(columns=drop_columns, inplace=True)
  df['review_datetime_utc'] = df['review_datetime_utc'].dt.date
  df.rename(columns={"name":"hotel_name",
                   "author_title": "user_name", "reviews": "num_of_reviews",
                   "author_id": "user_id", "review_text": "review", "review_rating": "rating",
                   "review_datetime_utc": "date"}, inplace=True)

  return df

ramada_la = etl(ramada_losangeles)
ramada_la["state"] = "CA"
ramada_la.head(3)

ramada_ny = etl(ramada_newyork)
ramada_ny["state"] = "NY"
ramada_ny.head(3)

El mismo hotel ramada tiene varios place_id

ramada_ny["place_id"] = "ChIJXfjqkJZZwokRTwca33qw-00"

ramada_ny.shape

ramada_ny.head(2)

ramada_or = etl(ramada_orlando)
ramada_or["state"] = "FL"
ramada_or.head(3)

ramada_re.shape

ramada_re = etl(ramada_reno)
ramada_re["state"] = "NV"
ramada_re.head(3)

df_list = [ramada_houston, ramada_ny, ramada_la, ramada_or, ramada_re]
df = pd.concat(df_list, ignore_index=True)

df.isna().mean()

df.dropna(subset="review", inplace=True)

users = pd.DataFrame()
users["user_id"] = df["user_id"]
users["user_name"] = df["user_name"]

users.isna().sum()

users["user_id"].fillna(1, inplace=True)

users.loc[users['user_id'] == 1, 'user_name'] = 'Desconocido'

users.drop_duplicates(inplace=True)

users

df.drop(columns="user_name", inplace=True)

df[df.duplicated(keep=False)].sort_values(by="review")

df.drop_duplicates(inplace=True)

hotel = pd.DataFrame()
hotel["hotel_name"] = df["hotel_name"]
hotel["place_id"] = df["place_id"]
hotel["state"] = df["state"]

hotel.drop_duplicates(inplace=True)

hotel

df.drop(columns=["review_id", "hotel_name", "state"], inplace=True)

df

users.to_csv("/content/drive/MyDrive/Colab Notebooks/hoteles-ramada/ramada-unido/users_ramada.csv", sep="|", index=False)

hotel.to_csv("/content/drive/MyDrive/Colab Notebooks/hoteles-ramada/ramada-unido/hotel_ramada.csv", sep="|", index=False)

df.to_csv("/content/drive/MyDrive/Colab Notebooks/hoteles-ramada/ramada-unido/reviews_ramada.csv", sep="|", index=False)
