==> notebooks/14. score_implementacion.ipynb <==
# Modelo Score de Sentimientos

## Preprocesado

!pip install demoji

import demoji
import re

def cleaning_text(text):
# TRATAR TRADUCCIONES EN REVIEWS, BORRA EL STRING "(ORIGINAL)" Y TODO A SU DERECHA
  keyword = "(Original)"
  keyword_index = text.find(keyword) # Encontrar la posición de la palabra clave
  if keyword_index != -1:
      text = text[:keyword_index + len(keyword)] # Eliminar todo después de la palabra clave

  text = text.replace("(Original)", "")
  text = text.replace("(Translated by Google)", "")

# TRATAR EMOJIS
  text = demoji.replace_with_desc(text, ' ') # reemplaza los emojis con palabras mas un espacio entre si

# BORRA SALTOS DE LINEA
  text = text.replace("\n", " ")
# BORRA CARACTERES NO ALFANUMERICOS
  text = re.sub(r'[^\w\s]', '', text)
# BORRA ESPACIOS AL PRINCIPIO Y AL FINAL Y CONVERTIR A MINUSCULAS
  text = text.strip()
  text = text.lower()
# DESPUES DE LIMPIADO VERIFICA SI HAY CARACTERES O NO, SI NO HAY LO CONVIERTE EN NULO
  if not text:
      return None
  return text

ejemplo = """(Translated by Google) Solo trip in May. Very attentive employees. I Can't complain. Clean, comfortable, big rooms! Air conditioning and excellent safe. Very nice and clean bathroom. Very good location.
Full breakfast, water dispenser and coffee all day.
I recommend it in price and quality! 😀💚

(Original)
Viaje sola en Mayo. Empleados muy atentos. Habitaciones limpias, cómodas, grandes! Aire acondicionado y caja fuerte excelente. Baño muy lindo y limpio. Muy buena ubicación.
Desayuno completo, dispenser de agua y cafe todo el día.
Lo recomiendo en precio y calidad!"""

cleaning_text(ejemplo)

### Ejemplo de uso

import pandas as pd

data = {'reviews': [
    "I love this hotel",
    "👍",
    "Amazing experience!",
    "...-()"
]}
df = pd.DataFrame(data)

# Usando apply
df["cleaned"] = df["reviews"].apply(cleaning_text)

# Usando comprehension lists (mas eficiente)
df["cleaned"] = [cleaning_text(x) for x in df["reviews"]]

df["cleaned"]

## Modelo

import nltk
nltk.download('vader_lexicon') # DESCARGAR DICCIONARIO UNA VEZ

from nltk.sentiment import SentimentIntensityAnalyzer

analyzer = SentimentIntensityAnalyzer()

def sentiment(text):
  text = cleaning_text(text) # LIMPIA EL TEXTO
  if not text: # SI EL TEXTO ESTA VACIO DESPUES DE LA LIMPIEZA, EL SCORE ES NEUTRO
    return 0
  else:   # SI NO ESTA VACIO, VADER LO CALCULA
    scores = analyzer.polarity_scores(text)
  return scores["compound"]

### Ejemplo de uso

import pandas as pd

data = {'reviews': [
    "I love this hotel",
    "👍",
    "Hated the food",
    "Amazing experience!",
    "...-()"
]}


df = pd.DataFrame(data)

# Usando comprehension lists (mas eficiente)
df["sentiment_score"] = [sentiment(x) for x in df["reviews"]]

df

# Para el EDA e Insights

Para el modelo no se incluyó la remoción de stopwords y lemmatización porque bajaban la precisión, pero puede ser util para la wordclouds en el EDA

## Stopwords

import nltk
from nltk.corpus import stopwords

nltk.download('stopwords') # DESCARGAR UNA VEZ
nltk.download('punkt')

stopwords_nltk = stopwords.words('english')

from nltk import word_tokenize

def remove_stopwords(text):
  tokens = word_tokenize(text) # TOKENIZA EL TEXTO EN PALABRAS SINGULARES
  filtered_sentence = [] # CREA UNA LISTA VACIA
  for i in tokens:
    if i not in stopwords_nltk: # SI LA PALABRA NO ESTA EN LA LISTA DE STOPWORDS, LA GUARDA EN LA LISTA VACIA
        filtered_sentence.append(i)

  return " ".join(filtered_sentence) # JUNTA LAS PALABRAS DE LA LISTA

remove_stopwords(ejemplo)

## Lemamtizacion

import spacy

nlp = spacy.load('en_core_web_sm') # DESCARGAR MODELO PREENTRENADO UNA VEZ

def lemmatize(text):
  doc = nlp(text)
  tokens = []
  for token in doc:
      tokens.append(token)
  lemmatization = [token.lemma_ for token in doc]
  sentence = " ".join(lemmatization)
  return sentence

lemmatize(ejemplo)

## Juntando las funciones

def cleaning_stopwords_lemmatization(text):
  text = cleaning_text(text)
  text = remove_stopwords(text)
  text = lemmatize(text)
  return text

cleaning_stopwords_lemmatization(ejemplo)
